# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Brgy.Tanod S.O.S Emergency Response API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin/self)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/approve:
    post:
      operationId: approveUser
      tags: [users]
      summary: Approve pending user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reject:
    post:
      operationId: rejectUser
      tags: [users]
      summary: Reject pending user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: User rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incident reports
      parameters:
        - name: tanodId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"

    post:
      operationId: createIncident
      tags: [incidents]
      summary: File an incident report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [resident, tanod, admin, superadmin]
        status:
          type: string
          enum: [pending, approved, rejected, active]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        barangayId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: ["string", "null"]
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
    RejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
        accuracy:
          type: ["number", "null"]
...
    Incident:
      type: object
      required: [id, tanodId, tanodName, type, status, timestamp]
      properties:
        id:
          type: string
        alertId:
          type: ["string", "null"]
        tanodId:
          type: string
        tanodName:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [pending, ongoing, resolved, referred]
        location:
          type: ["string", "null"]
        gpsLocation:
          oneOf:
            - $ref: "#/components/schemas/Location"
            - type: "null"
        description:
          type: string
        personsInvolved:
          type: ["string", "null"]
        actionsTaken:
          type: ["string", "null"]
        citizenName:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
    IncidentInput:
      type: object
      required: [type, description, status]
      properties:
        alertId:
          type: ["string", "null"]
        type:
          type: string
        status:
          type: string
        location:
          type: ["string", "null"]
        gpsLocation:
          oneOf:
            - $ref: "#/components/schemas/Location"
            - type: "null"
        description:
          type: string
        personsInvolved:
          type: ["string", "null"]
        actionsTaken:
          type: ["string", "null"]
        citizenName:
          type: ["string", "null"]
    IncidentUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
        actionsTaken:
          type: ["string", "null"]
        resolutionNotes:
          type: ["string", "null"]
...