# lib/api-zod/src/generated/types/incidentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brgy.Tanod S.O.S Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
import type { Location } from './location';

export interface IncidentInput {
  /** @nullable */
  alertId?: string | null;
  type: string;
  status: string;
  /** @nullable */
  location?: string | null;
  gpsLocation?: Location | null;
  description: string;
  /** @nullable */
  personsInvolved?: string | null;
  /** @nullable */
  actionsTaken?: string | null;
  /** @nullable */
  citizenName?: string | null;
}