# lib/api-zod/src/generated/types/rejectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Brgy.Tanod S.O.S Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface RejectInput {
  reason: string;
}